Fix issue to break out from infinite loop
The function `runParallelCommandsOnPods` had a `count` variable that was supposed to be exit the infinite loop when done but it wasn't incremented. I replace it with the `podNumber` and removed the check for the status code because we want to see the error message from container that we couldn't run command inside them such as evicted containers.